requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
dropbox==12.0.0
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
ply==3.11
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.29.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
stone==3.3.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
